// src/client/__generated__/types.ts
export interface LoginMutation_login_user {
  __typename: "User";
  name: string | null;
  id: string;
  email: string;
  role: UserRole;
}

export interface LoginMutation_login {
  __typename: "AuthPayload";
  user: LoginMutation_login_user;
  token: string;
}

export interface LoginMutation {
  login: LoginMutation_login;
}

export interface LoginMutationVariables {
  email: string;
  password: string;
}